[Radoub.Formats] fix: ERF ExtractResource uint overflow before bounds check (#2244)
ErfReader.ExtractResource compared `entry.Offset + entry.Size > erfBuffer.Length`
with both operands as uint. A crafted entry with Offset = 0xFFFFFFF0 and
Size = 0x20 wraps to 0x10, passes the bounds check, and triggers an OOB
Array.Copy.

Fix: promote sum to long before compare.

Tests: ExtractResource_OffsetPlusSizeWrapsUint_Throws now passes.
All 27 ErfReaderTests still pass.

+++ Radoub.Formats/Radoub.Formats/Erf/ErfReader.cs
@@ -318,7 +318,8 @@ public static class ErfReader
     /// </summary>
     public static byte[] ExtractResource(byte[] erfBuffer, ErfResourceEntry entry)
     {
-        if (entry.Offset + entry.Size > erfBuffer.Length)
+        // Promote to long to detect uint overflow before bounds check (#2244).
+        if ((long)entry.Offset + entry.Size > erfBuffer.Length)
             throw new InvalidDataException($"Resource '{entry.ResRef}' extends beyond file boundary");
 
         var data = new byte[entry.Size];
